libct/cg: support disable_cgroup_devices build tag
The idea is to use this tag for software which wants to use
libcontainer/cgroups packages, but does not require the functionality
to manage devices in cgroups, which brings some additional overhead,
especially in fs2 where it exports cilium/ebpf. Using this  build tag
enables that software to save over 900 Kb (text+data+bss), which
reduces runc binary size by more than 8%:

   text	   data	    bss	    dec	    hex	filename
7016344	3938048	 229704	11184096	 aaa7e0	runc.devices
6423082	3597960	 228808	10249850	 9c667a	runc.no-devices

Obviously, we do not want runc to be compiled with this tag, so add
a guard to main.go and libcontainer itself.

NOTE than the tag is used for fs, we still create a directory under
devices cgroup root, and it is used by Exists/GetPids/GetAllPids,
but otherwise nothing is done with devices. The effect is similar to
that of Resources.SkipDevices.

Document the tag in README, and test this tag in unittest target.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>