libct: prepareRoot: avoid extra bind mount
We need rootfs to be a mount point. To that effect, we bind mount it to
itself. In case the rootfs itself is a mount point, this is not needed.

Add the code to determine whether rootfs is a mount point (since we
already read mountinfo it does not incur any additional overhead),
and skip the bind mount in case rootfs is already a mount point.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>