Add support for CRIU's 'ignore' manage-cgroup mode
This adds an already existing CRIU cgroup mode to be selectable from
runc: 'ignore'.

This is a fix for errors seen with Podman containers that have been
restored multiple times into new cgroups. CRIU fails to checkpoint a
container that has already been restored once in a cgroup named
differently than the original container.

CRIU keeps information about the original cgroup and during restore in
a differently named cgroup CRIU still tries to restore to original
cgroup. This leads to a process with wrong cgroup settings and a second
checkpoint of that container fails.

With 'ignore' CRIU will ignore any cgroups and leave the cgroup setting
to runc.

Signed-off-by: Adrian Reber <areber@redhat.com>