libct: Don't use stale fds
The mountConfig fd field can be set by a previous iteration and we are
not clearing it.

Let's just move the mountConfig inside the loop, so we avoid all this
kinds of bugs in the future too. Note that none of the fields set on
mountConfig have secondary effects (they are just reading fields from
config/iConfig), so this is safe to do.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>