libct/cg/dev: initial support for Path
This brings in initial support for specifying a device access rule using
device path instead of (or in addition to) device type, major, and
minor.

In case the path is not specified, nothing is done.

Otherwise, it is checked that the path exists and is an actual block or
character device.

In case the path is specified together with type/major/minor,
a sanity check is performed to make sure that they match those
of the actual device.

In case the path is specified but type/major/minor isn't,
they are filled it from the actual device.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>