Bump go 1.20.2
Go 1.20.2 has an important fix to an issue described in [1].
Switch from using Go 1.19 from Dockerfile (used for release binaries and
some CI) as well as from cirrus-ci.

Note in cirrus-ci we have to specify the patch number explicitly, as
otherwise go 1.20 will be used, while in Dockerfile saying 1.20 means
to use 1.20.x (which is 1.20.2 now).

[1] https://github.com/golang/go/issues/58624

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>