libct/cg/sd/v2: rely on systemd to set device rules
It seems that the code added by commit b810da149008f1d7d0 had cgroup v1
in mind, where runc overwrites the rules set by systemd.

In cgroup v2, things are different, since both ebpf programs (the one
added by systemd and the one added by runc) have to say "allow" so that
a program can access a device. So, adding a second EBPF is not really
helpful (provided that systemd works as it should).

This commit does two things:

1. Removes adding a second EBPF when systemd is used.

2. Promotes debug-warnings to real warnings for cgroup v2 case,
   since for v2 case they are real problematic.

Initial version of this commit returned an error for cgroup v2 case, but
we've decided to demote it to a warning, otherwise we risk breaking
backward compatibility.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>