libct: document process.LogLevel field
This field used to hold a string representation of log level (like
"debug" or "info"). Since commit 6c4a3b13d132aa4 this is now a string
holding a numeric representation of log level (e.g. "4"). This was done
in preparation to commit f1b703fc45dd6142, and simplifies things in
a few places.

Let's document it.

Fixes: 6c4a3b13d132aa4
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>