MAINTAINERS: add Li Fu Bang
I am nominating @lifubang for the role of runc maintainer.

He provided a number of valuable contributions to runc, and demonstrated
both deep technical expertise and the long term commitment to the
project.

As noted in MAINTAINERS_GUIDE.md, we have a week to vote, and need to
get 66% of current maintainers' votes.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>