Support idmap mounts on volumes without userns.
Commit fda12ab1014e ("Support idmap mounts on volumes") introduces support for
idmap mount on volumes.
At this time, it required userns to be used and idmap mount must have the same
UID/GID mappings than userns.
This commit removes this requirement, so it is possible to use idmap mount
without userns.
You can, of course, use id map mount with userns, but the UID/GID mappings can
be different.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>