libcontainer: remove all mount logic from nsexec
With open_tree(OPEN_TREE_CLONE), it is possible to implement both the
id-mapped mounts and bind-mount source file descriptor logic entirely in
Go without requiring any complicated handling from nsexec.

This allows us to remove the amount of C code we have in nsexec, as well
as simplifying a whole host of places that were made more complicated
with the addition of id-mapped mounts and the bind sourcefd logic. The
one downside of this is that the bind sourcefd feature now depends on
Linux 5.4.

In addition, we can easily add support for id-mappings that don't match
the container's user namespace. The approach taken here is to use Go's
officially supported mechanism for spawning a process in a user
namespace, but (ab)use PTRACE_TRACEME to avoid actually having to exec a
different process. The most efficient way to implement this would be to
do clone() in cgo directly to run a function that just does
kill(getpid(), SIGSTOP) -- we can always switch to that if it turns out
this approach is too slow.

This also includes a partial fix for a bug in the handling of idmap
mounts and mount propagation. In short, the issue is that because we do
OPEN_TREE_CLONE on the host, the RootPropagation flag does not apply
(nor any other mount propagation flags configured in config.json) and
thus recursive bind-mounts can and will be leaked to the host. Because
this patch switches the feature from 9c444070ec7b ("Open bind mount
sources from the host userns") to use OPEN_TREE_CLONE, this resulted in
all bind-mounts having this behaviour. The partial fix here is to try to
emulate the behaviour of RootPropagation for bind-mounts from the host.

It turns out that bind-mounts inside containers were broken when using
the fd-passing feature from 9c444070ec7b ("Open bind mount sources from
the host userns") because the file descriptor opens were done before we
start doing any mounts in rootfs_linux.go. The solution for this is more
involved and is fixed in a separate patch. At the very least, this patch
doesn't worsen the mount propagation situation.

Fixes: fda12ab1014e ("Support idmap mounts on volumes")
Fixes: 9c444070ec7b ("Open bind mount sources from the host userns")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>